fix: clippy warnings — 0 errors, 0 warnings
- auth.rs: unnecessary parentheses in pattern match
- webhooks.rs: remove unnecessary i64 casts (3 places)
- applications.rs: redundant closure → method reference
- notifications.rs: useless format! → string literal (2 places)
- overage.rs: useless format! → string literal (3 places)
- admin.rs: unused function allow, unnecessary borrow
- billing/mod.rs: wildcard covers any other pattern
- billing/polar.rs: unused field prefix with underscore
- email.rs: Default derive + rename from_str → parse_lang
- email.rs: too_many_arguments allow (2 places)
- resend_email.rs: too_many_arguments allow
- worker/http.rs: manual char → array pattern
- worker/main.rs: unused const allow, too_many_arguments allow

+++ api/src/routes/webhooks.rs
@@ -246,7 +246,7 @@ async fn create_webhook(
             "UPDATE customers SET webhook_count = webhook_count + 1 WHERE id = $1 AND webhook_count < $2 RETURNING *",
         )
         .bind(customer.id)
-        .bind(customer.webhook_limit as i64)
+        .bind(customer.webhook_limit)
         .fetch_optional(&pool)
         .await?
     };
@@ -349,7 +349,7 @@ async fn batch_webhooks(
         )
         .bind(batch_count)
         .bind(customer.id)
-        .bind(customer.webhook_limit as i64)
+        .bind(customer.webhook_limit)
         .fetch_optional(&pool)
         .await?
     };
@@ -543,7 +543,7 @@ async fn replay_webhook(
             "UPDATE customers SET webhook_count = webhook_count + 1 WHERE id = $1 AND webhook_count < $2 RETURNING *",
         )
         .bind(customer.id)
-        .bind(customer.webhook_limit as i64)
+        .bind(customer.webhook_limit)
         .fetch_optional(&pool)
         .await?
     };
@@ -681,7 +681,7 @@ async fn batch_replay(
                 "UPDATE customers SET webhook_count = webhook_count + 1 WHERE id = $1 AND webhook_count < $2 RETURNING *",
             )
             .bind(customer.id)
-            .bind(customer.webhook_limit as i64)
+            .bind(customer.webhook_limit)
             .fetch_optional(&pool)
             .await?
         };
